+++ src/pages/_document.js
@@ -1,4 +1,5 @@
 import Document, { Html, Head, Main, NextScript } from 'next/document'
+import { ServerStyleSheet } from 'styled-components';
 
 class MyDocument extends Document {
   render() {
@@ -24,6 +25,32 @@ class MyDocument extends Document {
       </Html>
     )
   }
+}
+
+MyDocument.getInitialProps = async (ctx) => {
+    const sheet = new ServerStyleSheet()
+    const originalRenderPage = ctx.renderPage
+
+    try {
+        ctx.renderPage = () =>
+            originalRenderPage({
+                enhanceApp: (App) => (props) =>
+                    sheet.collectStyles(<App {...props} />),
+            })
+
+        const initialProps = await Document.getInitialProps(ctx)
+        return {
+            ...initialProps,
+            styles: (
+                <>
+                    {initialProps.styles}
+                    {sheet.getStyleElement()}
+                </>
+            ),
+        }
+    } finally {
+        sheet.seal()
+    }
 }
 
 export default MyDocument
