+++ src/styles/variables.css
@@ -26,6 +26,7 @@
     --input-color: hsl(var(--hue-color), 70%, 96%);
     --body-color: hsl(var(--hue-color), 60%, 99%);
     --container-color: #FFF;
+    --white-color: #FFF;
     /*========== Font and typography ==========*/
     /* --body-font: 'Poppins', sans-serif; */
     /* --icons-semantic-font: Icons; */
